# Fix half tensor printing plus speedup large tensor printing