@vadimkantorov 2018-11-28
@fmassa It seems that a giant GPU tensor print timing is not in the tests. Could you please add such a test? In my example CPU tensor printing with 1-sized first dim is slow (0.5 sec), but still much faster than GPU printing (40 sec).