@fmassa 2018-11-28
@vadimkantorov I'd rather not have a giant GPU tensor in tests as that might slowdown tests and make them use much more memory.
But we do have a test that verifies that the summarized tensor is of the right shape, so that should cover part of the problem.

@vishwakftw yes, this should fix #12093 as well